src/train.py
import hydra
import pyrootutils
from omegaconf import DictConfig

root = pyrootutils.setup_root(
    search_from=__file__,
    indicator=[".git", "pyproject.toml"],
    pythonpath=True,
    dotenv=True,
)
_HYDRA_PARAMS = {
    "version_base": "1.3",
    "config_path": str(root / "configs"),
    "config_name": "train.yaml",
}


@hydra.main(**_HYDRA_PARAMS)
def main(cfg: DictConfig) -> None:
    pass


if __name__ == "__main__":
    main(None)
